requirements.txt
Django==1.8.7
django-jinja==2.1.1
django-redis==4.3.0
docutils==0.12
Jinja2==2.8
MarkupSafe==0.23
# mysqlclient==1.3.7
PyMySQL==0.7.1
pypinyin==0.9.3
pytz==2016.3
redis==2.10.5
requests==2.9.0
django-configurations==1.0
gunicorn==19.4.5
user-agents==1.0.1
gevent==1.1rc1
django-debug-toolbar==1.4
aiohttp==0.21.2
